extracted deleteIndex() into ElasticSearchIOTestUtils to be used in integration tests
